Alteration to SQL Custom Field - Display and Form Field Rendering
This alteration is to allow for the situation where an sql SELECT needs to contain a dynamically computed value.

As an example:
SELECT id AS value, title AS text FROM #__content WHERE catid = 8 AND created_by = %%userid%% ORDER BY title ASC

the WHERE clause has created_by = a token %%userid%%

the token can be replaced at runtime with the LOGGED ON user id.

In this simple use case, the existing behaviour of the custom field is not altered, and only in the situation where the field developer needs to make use of the computed UserID, does the change kick in.